+++ hpowbcli/src/main/java/org/monarchinitiative/hpoworkbench/cmd/GitCommand.java
@@ -1,5 +1,7 @@
 package org.monarchinitiative.hpoworkbench.cmd;
 
+import com.beust.jcommander.Parameter;
+import com.beust.jcommander.Parameters;
 import org.apache.log4j.Logger;
 import org.monarchinitiative.hpoworkbench.github.GitHubIssue;
 import org.monarchinitiative.hpoworkbench.github.GitHubIssueRetriever;
@@ -8,19 +10,19 @@ import org.monarchinitiative.hpoworkbench.word.GitIssue2Doc4J;
 
 import java.util.List;
 
+
+@Parameters(commandDescription = "batch. Post a batch of GitHub issues to the HPO tracker.")
 public class GitCommand  extends HPOCommand {
     private static final Logger LOGGER = Logger.getLogger(GitCommand.class.getName());
-
-    private final String issueLabel;
+    @Parameter(names={"-l","--label"},required = true,description = "git issue label")
+    private  String issueLabel;
 
     /**
      * Create a word document with up to 30 open issues for the label. This is intended to be used
      * to make a summary of open documents for collaborators but unfortunately is limited to up to
      * 30 GitHub issues.
-     * @param label GitHub label
      */
-    public GitCommand(String label) {
-        issueLabel=label;
+    public GitCommand() {
     }
 
     public void run() {